From float to double precision

+++ bbttAnalysis/src/HHbbttSelectorAlg.h
@@ -275,7 +275,7 @@ private:
     std::vector<HHBBTT::Booleans> m_inputCutKeys;
     Gaudi::Property<bool> m_saveCutFlow{this, "saveCutFlow", false};
     long long int m_total_events{0};
-    float m_total_mcEventWeight{0.f};
+    double m_total_mcEventWeight{0.0};
     CP::SysReadDecorHandle<float>
       m_generatorWeight{ this, "generatorWeight", "generatorWeight_%SYS%", "MC event weights" };
